requirements.txt

aiocache==0.12.3
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dependency-injector==4.42.0
ecdsa==0.19.0
Faker==30.8.2
fastapi==0.115.0
fastapi-utils==0.7.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
loguru==0.7.2
Mako==1.3.5
MarkupSafe==3.0.0
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psutil==5.9.8
psycopg2-binary==2.9.10
pyaml-env==1.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.31.0
win32-setctime==1.1.0